swal added'

+++ src/Pages/Shared/DashBoard/ManageOrders/ManageOrderRaw.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import swal from 'sweetalert';
 
 const ManageOrderRaw = ({order, index}) => {
     const { productName, productImg, totalPrice, orderQuantity, _id, paid } = order
@@ -10,6 +11,7 @@ const ManageOrderRaw = ({order, index}) => {
         const payment={
             status:currentStatus
         }
+
         fetch(`https://hidden-sea-29105.herokuapp.com/paymentOrder/${_id}`,{
             method:'PUT',
             headers:{
@@ -31,13 +33,27 @@ const ManageOrderRaw = ({order, index}) => {
 
 
     const handelDelete = (id) => {
-        fetch(`https://hidden-sea-29105.herokuapp.com/orders/${id}`, {
-            method: 'DELETE',
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
-            })
+        swal({
+            title: "Are You Sure to Delete?",
+            text: `Item Name: ${productName}`,
+            icon: "error",
+            buttons: ["Cancel", "Delete"],
+            closeOnClickOutside: false,
+          })
+          .then((willPay) => {
+            if (willPay) {
+                swal("Item Deleted!", "Successfully Item Deleted!", "success")
+                const url= `https://hidden-sea-29105.herokuapp.com/orders/${id}`
+                fetch(url,{
+                  method:'DELETE'
+              })
+              .then(res=>res.json())
+              .then(data=>{
+                  if(data.deletedCount>0){
+                 }
+             })
+            } 
+          });
     }
     return (
         <tr>
